rename global-state-extras
